+++ db4e/constants/DLabel.py
@@ -37,7 +37,7 @@ class DLabel:
     DISABLE: Final[str] = "Disable"
     DONATIONS: Final[str] = "Donations"
     DONATIONS_WALLET: Final[str] = "Donations Wallet"
-    ENABLED: Final[str] = "Is Enabled"
+    ENABLED: Final[str] = "Is Running"
     ELEMENT_TYPE: Final[str] = "Component"
     ENABLED_FLAG: Final[str] = "Enabled flag"
     END: Final[str] = "End"